chore: create ColorSpaceValue component
